+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "msn-line-tool-v0.10.4";
+const CACHE_NAME = "msn-line-tool-v0.10.5";
 const APP_SHELL = [
   "./",
   "./index.html",